Solved problem with displaying edges. Also, moved the map example out of the code.
Now the map is stored in a .txt file that has special format to it. Two
methods from Sector.java are separated into a standalone Geometry class
in order to simplify the code. Also test cases added for Edge. java.
Some comments added to and a bit of minor refactoring (e.g. Most of the
outer interface is moved from Main.java to Map.java so that Main.java
will be in future able to communicate with other classes and is not
completely bound to the map). Moving the player along the map is
currently not working which is due to the ongoing fixing of the
sectorChange method in Sector.java.